requirements.txt
blinker==1.9.0
click==8.3.1
exceptiongroup==1.3.1
Flask==3.1.2
Frozen-Flask==1.0.2
greenlet==3.3.0
gunicorn==23.0.0
html5lib==1.1
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
pycodestyle==2.14.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
six==1.17.0
SQLAlchemy==2.0.44
tomli==2.3.0
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.1.4
WTForms==3.2.1
